requirements.txt
asgiref==3.2.10
coverage==5.3
Django==3.1.1
djangorestframework == 3.12.2
requests == 2.25.1
psycopg2-binary==2.8.6
pytz==2020.1
sqlparse==0.3.1
